requirements.txt
edapi
slack_sdk
python-dotenv
html-slacker
flask
psutil
